web-animations/animation-timeline/animation-timeline-play.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationTimeline.play() method tests</title>
<meta name="assert" content="Creates a new AnimationPlayer object associated with this timeline that begins playback as soon as it is ready. The new AnimationPlayer object is created using the the AnimationPlayer constructor passing this AnimationTimeline object as the timeline parameter and source as the source parameter. Following construction of the AnimationPlayer object, the procedure to play a player is performed on the newly constructed object">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationtimeline-play">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
var TIMEOUT = ANIMATION_DURATION / 20;

test(function() {
    var animation = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var player = document.timeline.play(animation);

    assert_not_equals(player, null, 'Player created by AnimationTimeline.play() ' +
        'method should be not null');
}, 'Player created by AnimationTimeline.play() method is not null');

test(function() {
    var player = document.timeline.play(null);

    assert_not_equals(player, null, 'Player created by AnimationTimeline.play(null) ' +
        'method should be not null');
}, 'Player created by AnimationTimeline.play(null) method is not null');

test(function() {
    var player = document.timeline.play();

    assert_not_equals(player, null, 'Player created by AnimationTimeline.play() ' +
        'method without arguments should not be null');
}, 'Player created by AnimationTimeline.play() method without arguments arguments is not null');

test(function() {
    var animation = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var player = document.timeline.play(animation);

    assert_class_string(player, 'AnimationPlayer', 'Player created by ' +
        'AnimationTimeline.play() method should be AnimationPlayer object instance');
}, 'Test AnimationTimeline.play() method creates a new AnimationPlayer object instance');

test(function() {
    var player = document.timeline.play(null);

    assert_class_string(player, 'AnimationPlayer', 'Player created by ' +
        'AnimationTimeline.play(null) method should be AnimationPlayer object instance');
}, 'Test AnimationTimeline.play(null) method creates a new AnimationPlayer object instance');

test(function() {
    var player = document.timeline.play();

    assert_class_string(player, 'AnimationPlayer', 'Player created by ' +
        'AnimationTimeline.play() method without arguments ' +
        'should be AnimationPlayer object instance');
}, 'Test AnimationTimeline.play() method without arguments creates ' +
    'a new AnimationPlayer object instance');

test(function() {
    var animation = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var player = document.timeline.play(animation);

    assert_equals(player.source, animation, 'The source attribute of the created ' +
        'AnimationPlayer should be set to source parameter value');
}, 'Test that the source attribute of the created AnimationPlayer is set to source ' +
    'parameter value');

test(function() {
    var player = document.timeline.play(null);

    assert_equals(player.source, null, 'The source attribute of the created ' +
        'AnimationPlayer should be null if source parameter value is null');
}, 'Test that the source attribute of the created AnimationPlayer is null if source ' +
    'parameter value is null');

test(function() {
    var player = document.timeline.play();

    assert_equals(player.source, null, 'The source attribute of the created ' +
        'AnimationPlayer should be null if source parameter value is not specified');
}, 'Test that the source attribute of the created AnimationPlayer is null if source ' +
    'parameter value is not specified');

test(function() {
    var animation = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var player = document.timeline.play(animation);

    assert_equals(player.timeline, document.timeline, 'AnimationPlayer, ' +
        'created by play() method should be accociated with this timeline');
}, 'Test AnimationTimeline.play() method accociates returned AnimationPlayer ' +
    'with timeline instance used to call method play()');

test(function() {
    var doc = newHTMLDocument();
    var animation = new Animation(createDiv(this, doc), KEYFRAMES, ANIMATION_DURATION);
    var player = doc.timeline.play(animation);

    assert_equals(player.timeline, doc.timeline, 'AnimationTimeline.play() should ' +
        'associate AnimationPlayer with timeline instance used to call play() method');
}, 'Test AnimationTimeline.play() method accociates returned AnimationPlayer ' +
    'with timeline instance used to call method play(). Using timeline of another document');

async_test(function(t) {
    var target = createDiv(t);
    var player = document.timeline.play(new Animation(target, KEYFRAMES, ANIMATION_DURATION));

    player.ready.then(t.step_func(function() {
        setTimeout(t.step_func(function() {
            assert_approx_equals(player.currentTime, TIMEOUT, TIME_EPSILON, 'AnimationTimeline.play() ' +
                'should run play a player procedure');
            assert_equals(player.playState, 'running', 'Play state should be in running state');
            assert_approx_equals(window.getComputedStyle(target).top,
                getExpectedTop(TIMEOUT), COMPUTED_STYLE_EPSILON,
                'Animation effect should be running');
            t.done();
        }), TIMEOUT);
    }), t.unreached_func('Player ready promise should not be rejected'));
}, 'Test AnimationTimeline.play() runs play a player procedure');
</script>
</body>
